requirements.txt
fastapi==0.119.1
langchain==1.0.1
langchain-google-genai==3.0.0
langchain-huggingface==1.0.0
numpy==2.3.4
python-dotenv==1.1.1
python-multipart==0.0.20
uvicorn==0.38.0